refactor: change Survey conversions to extension functions
