+++ sharedTest/src/main/java/com/sharedtest/persistence/local/LocalDataStoreHelper.kt
@@ -16,9 +16,9 @@
 package com.sharedtest.persistence.local
 
 import com.google.android.ground.model.Survey
+import com.google.android.ground.persistence.local.room.converter.toSurveyEntity
 import com.google.android.ground.persistence.local.room.dao.SurveyDao
 import com.google.android.ground.persistence.local.room.dao.insertOrUpdate
-import com.google.android.ground.persistence.local.room.entity.SurveyEntity
 import io.reactivex.Completable
 import javax.inject.Inject
 
@@ -29,6 +29,6 @@ class LocalDataStoreHelper @Inject constructor(private val surveyDao: SurveyDao)
   }
 
   fun insertSurvey(survey: Survey) {
-    surveyDao.insertOrUpdate(SurveyEntity.fromSurvey(survey)).commit()
+    surveyDao.insertOrUpdate(survey.toSurveyEntity()).commit()
   }
 }
